refactor(inter-protocol): factor out startEconCommittee.js
... in order to avoid importing sim-behaviors.js into boot-psm.js
