+++ packages/inter-protocol/src/proposals/startEconCommittee.js
@@ -1,0 +1,90 @@
+// @ts-check
+import {
+  assertPathSegment,
+  makeStorageNodeChild,
+} from '@agoric/vats/src/lib-chainStorage.js';
+import { E } from '@endo/far';
+import { makeTracer } from '../makeTracer.js';
+
+const trace = makeTracer('RunEconBehaviors', false);
+
+/** @type {(name: string) => string} */
+const sanitizePathSegment = name => {
+  const candidate = name.replace(/[ ,]/g, '_');
+  assertPathSegment(candidate);
+  return candidate;
+};
+
+/**
+ * @typedef {object} EconCommitteeOptions
+ * @property {string} [committeeName]
+ * @property {number} [committeeSize]
+ */
+
+/**
+ * @param {import('./econ-behaviors').EconomyBootstrapPowers} powers
+ * @param {object} [config]
+ * @param {object} [config.options]
+ * @param {EconCommitteeOptions} [config.options.econCommitteeOptions]
+ */
+export const startEconomicCommittee = async (
+  {
+    consume: { board, chainStorage, zoe },
+    produce: { economicCommitteeCreatorFacet },
+    installation: {
+      consume: { committee },
+    },
+    instance: {
+      produce: { economicCommittee },
+    },
+  },
+  { options: { econCommitteeOptions = {} } = {} },
+) => {
+  const COMMITTEES_ROOT = 'committees';
+  trace('startEconomicCommittee');
+  const {
+    committeeName = 'Initial Economic Committee',
+    committeeSize = 3,
+    ...rest
+  } = econCommitteeOptions;
+
+  const committeesNode = await makeStorageNodeChild(
+    chainStorage,
+    COMMITTEES_ROOT,
+  );
+  const storageNode = await E(committeesNode).makeChildNode(
+    sanitizePathSegment(committeeName),
+  );
+  const marshaller = await E(board).getReadonlyMarshaller();
+
+  const { creatorFacet, instance } = await E(zoe).startInstance(
+    committee,
+    {},
+    { committeeName, committeeSize, ...rest },
+    {
+      storageNode,
+      marshaller,
+    },
+  );
+
+  economicCommitteeCreatorFacet.resolve(creatorFacet);
+  economicCommittee.resolve(instance);
+};
+harden(startEconomicCommittee);
+
+export const ECON_COMMITTEE_MANIFEST = harden({
+  [startEconomicCommittee.name]: {
+    consume: {
+      board: true,
+      chainStorage: true,
+      zoe: true,
+    },
+    produce: { economicCommitteeCreatorFacet: 'economicCommittee' },
+    installation: {
+      consume: { committee: 'zoe' },
+    },
+    instance: {
+      produce: { economicCommittee: 'economicCommittee' },
+    },
+  },
+});
