+++ packages/inter-protocol/test/amm/vpool-xyk-amm/setup.js
@@ -17,8 +17,8 @@ import { makeTracer } from '../../../src/makeTracer.js';
 import {
   setupAmm,
   setupReserve,
-  startEconomicCommittee,
 } from '../../../src/proposals/econ-behaviors.js';
+import { startEconomicCommittee } from '../../../src/proposals/startEconCommittee.js';
 import {
   installGovernance,
   makeMockChainStorageRoot,
